BUG FIX UNDO: delete and create new intent and delete create connector

+++ src/app/chatbot-design-studio/services/intent.service.ts
@@ -350,7 +350,7 @@ export class IntentService {
   public async deleteFaq(intent: Intent): Promise<boolean> { 
     console.log('[INTENT SERVICE] -> deleteFaq, ');
     return new Promise((resolve, reject) => {
-      this.faqService.deleteFaq(intent.id, intent.intent_id).subscribe((data) => {
+      this.faqService.deleteFaq(intent.id, intent.intent_id, intent.id_faq_kb).subscribe((data) => {
         this.prevListOfIntent = JSON.parse(JSON.stringify(this.listOfIntents));
         resolve(true);
       }, (error) => {
@@ -384,7 +384,7 @@ export class IntentService {
       let intentsToUpdateUndo = [intentPrev];
       let intentsToUpdateRedo = [intentNow];
       if(!connector)this.addIntentToUndoRedo('PUT', intentPrev, intentNow, intentsToUpdateUndo, intentsToUpdateRedo);
-      else if(connector) this.addIntentToUndoRedo('CONN', intentPrev, intentNow, intentsToUpdateUndo, intentsToUpdateRedo, connector);
+      else if(connector && connector.undo) this.addIntentToUndoRedo('CONN', intentPrev, intentNow, intentsToUpdateUndo, intentsToUpdateRedo, connector);
     }
     if(!timeout)timeout = 0;
     return new Promise((resolve, reject) => {
@@ -460,7 +460,7 @@ export class IntentService {
       this.connectorService.deleteConnectorsFromActionByActionId(actionId);
       console.log('[CDS-INTENT-SERVICES] aggiorno intent di partenza', intentToUpdate);
       // const responseIntent = this.updateIntent(intentToUpdate);
-      const responseIntent = this.onUpdateIntentWithTimeout(intentToUpdate);
+      const responseIntent = this.onUpdateIntentWithTimeout(intentToUpdate, 0, true);
       if(responseIntent){
         // console.log('[CDS-INTENT-SERVICES] update Intent: OK');
         this.behaviorIntent.next(intentToUpdate);
@@ -541,18 +541,18 @@ export class IntentService {
   // START ACTION FUNCTIONS //
 
   /** update title of intent */
-  public changeIntentName(intent){
-    setTimeout(async () => {
+  public async changeIntentName(intent){
+    // setTimeout(async () => {
       let prevIntent = this.prevListOfIntent.find((obj) => obj.intent_id === intent.intent_id);
       if(intent.intent_display_name !== prevIntent.intent_display_name){
-        const response = await this.onUpdateIntentWithTimeout(intent, 500);
+        const response = await this.onUpdateIntentWithTimeout(intent, 0, true);
         if(response){
           // this.behaviorIntents.next(this.listOfIntents);
           // this.refreshIntent(intent);
           this.setDragAndListnerEventToElement(intent.intent_id);
         }
       }
-    }, 500);
+    // }, 500);
   }
 
   // moving new action in intent from panel elements
@@ -565,15 +565,16 @@ export class IntentService {
     currentIntent.actions.splice(event.currentIndex, 0, newAction);
     this.behaviorIntent.next(currentIntent);
     this.connectorService.updateConnector(currentIntent.intent_id);
-    setTimeout(async () => {
-      const responseCurrentIntent = await this.onUpdateIntentWithTimeout(currentIntent);
-      if(responseCurrentIntent){
-        // const fromEle = document.getElementById(currentIntent.intent_id);
-        // this.connectorService.updateConnector(currentIntent.intent_id);
-        console.log('update current Intent: OK');
-        //this.behaviorIntent.next(currentIntent);
-      }
-    }, 0);
+    this.onUpdateIntentWithTimeout(currentIntent, 0, true);
+    // setTimeout(async () => {
+      // const responseCurrentIntent = await this.onUpdateIntentWithTimeout(currentIntent, 0, true);
+      // if(responseCurrentIntent){
+      //   // const fromEle = document.getElementById(currentIntent.intent_id);
+      //   // this.connectorService.updateConnector(currentIntent.intent_id);
+      //   console.log('update current Intent: OK');
+      //   //this.behaviorIntent.next(currentIntent);
+      // }
+    // }, 0);
     return newAction
   }
 
@@ -619,8 +620,39 @@ export class IntentService {
    */
   public deleteIntentToListOfIntents(intentId){
     // console.log("[CDS-INTENT-SERVICES] elimino l'intent alla lista di intent", intentId);
+    // devo aggiornare tutti gli intent connessi
+    // this.deleteAllConnectorsInAndUpdateIntents();
     this.listOfIntents = this.listOfIntents.filter((intent: any) => intent.intent_id !== intentId);
   }
+
+  public updateAndSaveAllIntentsConnectedToDeletedIntent(intentId){
+    // console.log("[CDS-INTENT-SERVICES] updateAllIntentsConnectedToDeletedIntent: intentId ", intentId);
+    let arrayOfIntents = [];
+    const listConnectors = this.connectorService.searchConnectorsInOfIntent(intentId);
+    listConnectors.forEach(element => {
+      const splitFromId = element.fromId.split('/');
+      const intentToUpdateId = splitFromId[0];
+      const resp = arrayOfIntents.some((obj) => obj.intent_id === intentToUpdateId);
+      if(!resp){
+        let intent = this.listOfIntents.find((intent: any) => intent.intent_id === intentToUpdateId);
+        arrayOfIntents.push(intent);
+      }
+    });
+    arrayOfIntents.forEach(intent => {
+      // cerca in tutto l'array di actions se c'è una action che contiene intentId (quindi connessa con l'intent da eliminare) e in questo caso assegna il campo a null
+      for (let i = 0; i < intent.actions.length; i++) {
+        const object = intent.actions[i];
+        for (const key in object) {
+          // console.log("[CDS-INTENT-SERVICES] object[key]: ", object[key], intentId);
+          if (typeof object[key] === 'string' && object[key].includes(intentId)) {
+            object[key] = null;
+          }
+        }
+      }
+      this.updateIntent(intent);
+    });
+  }
+
 
   /** getListOfActions */
   public getListOfActions(){
@@ -707,7 +739,7 @@ export class IntentService {
       this.connectorService.updateConnector(intentToUpdate.intent_id);
       this.controllerService.closeAllPanels();
       // this.connectorService.deleteConnectorsFromActionByActionId(this.actionSelectedID);
-      const responseIntent = this.onUpdateIntentWithTimeout(intentToUpdate);
+      const responseIntent = this.onUpdateIntentWithTimeout(intentToUpdate, 0, true);
       if(responseIntent){
         // this.connectorService.movedConnector(intentToUpdate.intent_id);
         console.log('update Intent: OK');
@@ -933,7 +965,7 @@ export class IntentService {
     if(!pos || pos<0)pos = this.listOfIntents.length;
     let UNDO = this.setUndoRedoObject(type, typeUNDO, typeREDO, pos, intentPre, intentNow, intentsToUpdateUndo, intentsToUpdateRedo, connector);
     this.arrayUNDO.push(UNDO);
-    console.log('[INTENT SERVICE] ***** AGGIUNGO ', UNDO, this.arrayUNDO );
+    console.log('[INTENT UNDO] ***** AGGIUNGO ', UNDO, this.arrayUNDO );
     // this.arrayUNDO = this.arrayUNDO.slice(10);
     this.lastActionUndoRedo = false;
     this.arrayREDO = [];
@@ -954,7 +986,7 @@ export class IntentService {
    * @returns 
    */
   private setUndoRedoObject(type, typeUNDO, typeREDO, pos, intentPre, intentNow, intentsToUpdateUndo, intentsToUpdateRedo, connector?){
-    console.log('[INTENT SERVICE] -> setUndoRedoObject ',intentNow, intentPre, intentsToUpdateUndo, intentsToUpdateRedo, connector);
+    console.log('[INTENT UNDO] -> setUndoRedoObject ',intentNow, intentPre, intentsToUpdateUndo, intentsToUpdateRedo, connector);
     let obj = {
       type: type,
       pos: pos,
@@ -984,7 +1016,7 @@ export class IntentService {
       this.arrayREDO.push(objUNDO);
       // console.log('[INTENT SERVICE] -> RESTORE UNDO: ', this.arrayREDO);
       this.restoreIntent(objUNDO.pos, objUNDO.undo);
-      console.log('[INTENT SERVICE] -> ho aggiornato gli array dopo UNDO ', this.arrayUNDO, this.arrayREDO);
+      console.log('[INTENT UNDO] -> ho aggiornato gli array dopo UNDO ', this.arrayUNDO, this.arrayREDO);
     }
   }
 
@@ -997,7 +1029,7 @@ export class IntentService {
       this.arrayUNDO.push(objREDO);
       // console.log('[INTENT SERVICE] -> RESTORE REDO: ', objREDO);
       this.restoreIntent(objREDO.pos, objREDO.redo);
-      console.log('[INTENT SERVICE] -> ho aggiornato gli array dopo REDO ', this.arrayUNDO, this.arrayREDO);
+      console.log('[INTENT UNDO] -> ho aggiornato gli array dopo REDO ', this.arrayUNDO, this.arrayREDO);
     }
   }
 
@@ -1014,7 +1046,7 @@ export class IntentService {
     const intentsToUpdate = object.intentsToUpdate;
     const connector = object.connector;
     this.lastActionUndoRedo = intent.intent_id;
-    console.log('[INTENT SERVICE] -> restoreAction : typeAction:', object, intentsToUpdate);
+    console.log('[INTENT UNDO] -> restoreAction : typeAction:', object, intentsToUpdate);
     if(restoreAction == 'DEL'){
       this.restoreDel(intent, intentsToUpdate);
       return;
@@ -1058,7 +1090,7 @@ export class IntentService {
     if(isOnTheStage){
       this.connectorService.createConnectorsOfIntent(intent);
       intentsToUpdate.forEach(element => {
-        console.log('[INTENT SERVICE] -> REPLACE ', element);
+        console.log('[INTENT UNDO] -> REPLACE ', element);
         this.listOfIntents = this.replaceIntent(element, this.listOfIntents);
         // this.refreshIntent(element); 
         setTimeout(()=> {
@@ -1096,7 +1128,7 @@ export class IntentService {
     this.deleteIntentToListOfIntents(intent.intent_id);
     this.connectorService.deleteConnectorsOfBlock(intent.intent_id); // cancello tutti i connettori IN e OUT
     intentsToUpdate.forEach(element => {
-      console.log('[INTENT SERVICE] -> REPLACE ', element);
+      console.log('[INTENT UNDO] -> REPLACE ', element);
       this.listOfIntents = this.replaceIntent(element, this.listOfIntents);
       setTimeout(()=> {
         // ATTENZIONE!!! le seguenti azioni devono essere svolte solo dopo che l'elemento è stato aggiornato (gli attributi __ dovrebbero essere generati quando carichiamo i dati dal server!!!)
@@ -1127,7 +1159,7 @@ export class IntentService {
    */
   private async restorePut(intent, intentsToUpdate){
     this.listOfIntents = this.replaceIntent(intent, this.listOfIntents);
-    console.log('[INTENT SERVICE] -> REPLACE INTENT: ', intent);
+    console.log('[INTENT UNDO] -> REPLACE INTENT: ', intent);
     // this.refreshIntent(intent); // aggiorno gli attributi custom ex: __isConnected !!! DA RIFATTORIZZARE !!!
     setTimeout(()=> {
       this.connectorService.deleteConnectorsOutOfBlock(intent.intent_id); // false, false
@@ -1138,15 +1170,17 @@ export class IntentService {
   
     intentsToUpdate.forEach(element => {
       this.listOfIntents = this.replaceIntent(element, this.listOfIntents);
-      console.log('[INTENT SERVICE] -> REPLACE ELEMENT: ', element);
+      console.log('[INTENT UNDO] -> REPLACE ELEMENT: ', element);
       // this.refreshIntent(element);
-      setTimeout(()=> {
-        this.connectorService.deleteConnectorsOutOfBlock(intent.intent_id); // false, false
-        this.connectorService.deleteConnectorsBrokenOutOfBlock(element.intent_id); // false, false
-        this.connectorService.createConnectorsOfIntent(element); // false, false
-        this.connectorService.updateConnector(element.intent_id);
-      }, 100);
-      this.updateIntent(element); // async
+      if( element.intent_id !== intent.intent_id){
+        setTimeout(()=> {
+          this.connectorService.deleteConnectorsOutOfBlock(intent.intent_id); // false, false
+          this.connectorService.deleteConnectorsBrokenOutOfBlock(element.intent_id); // false, false
+          this.connectorService.createConnectorsOfIntent(element); // false, false
+          this.connectorService.updateConnector(element.intent_id);
+        }, 100);
+        this.updateIntent(element); // async
+      }
     });
     this.updateIntent(intent); // async
     this.refreshIntents();
@@ -1167,18 +1201,18 @@ export class IntentService {
    * refreshIntents: fa scattare l'evento e aggiorna l'elenco degli intents (listOfIntents) in tutti i componenti sottoscritti, come cds-panel-intent-list 
    */
   private async restoreConnector(intent, connector){
-    
-    // setTimeout(()=> {
+    this.updateIntent(intent); // async
+    this.listOfIntents = this.replaceIntent(intent, this.listOfIntents); /// Arggggg!! questa mmerda ricarica tutte le action e mi svuota tutti i connettori true false
+ 
+    setTimeout(()=> {
         const arrow = document.getElementById(connector.id);
-        console.log('[INTENT SERVICE] -> arrow: ', arrow, connector);
+        console.log('[INTENT UNDO] -> arrow: ', arrow, connector);
         if(arrow)this.connectorService.deleteConnector(connector.id);
         else this.connectorService.createNewConnector(connector.fromId, connector.toId);
       // this.connectorService.updateConnector(intent.intent_id, false);
-    // }, 100);
+    }, 100);
 
-    // this.listOfIntents = this.replaceIntent(intent, this.listOfIntents); /// Arggggg!! questa mmerda ricarica tutte le action e mi svuota tutti i connettori true false
-    this.updateIntent(intent); // async
-    // this.refreshIntent(intent);
+       // this.refreshIntent(intent);
     return;
   }
   /************************************************/
